fix: remove view bids option when no bids
