+++ src/containers/Tables/OwnedDomainsTable/OwnedDomainsTable.constants.ts
@@ -1,6 +1,3 @@
-import settingsIcon from './assets/settings.svg';
-import dollarSignIcon from './assets/dollar-sign.svg';
-
 export const HEADERS = [
 	{
 		label: 'Domain',
@@ -27,14 +24,3 @@ export const ACTION_KEYS = {
 	SETTINGS: 'Domain Settings',
 	VIEW_BIDS: 'View Bids',
 };
-
-export const ACTIONS = [
-	{
-		icon: settingsIcon,
-		title: ACTION_KEYS.SETTINGS,
-	},
-	{
-		icon: dollarSignIcon,
-		title: ACTION_KEYS.VIEW_BIDS,
-	},
-];
